refactor: move gemini config to team.po.config package #52
