+++ src/main/java/team/po/config/GeminiProperties.java
@@ -1,4 +1,4 @@
-package team.po.feature.devguide.client;
+package team.po.config;
 
 import org.springframework.boot.context.properties.ConfigurationProperties;
